requirements.txt
# Core Dependencies
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision>=1.2.0

# Image Processing
Pillow>=10.0.0
scikit-image>=0.21.0
opencv-python>=4.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & Metrics
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.66.0